add a workaround for mobile metamask
