+++ dapp/src/components/buySell/BuySellWidget.js
@@ -19,6 +19,7 @@ import { sleep } from 'utils/utils'
 import { providersNotAutoDetectingOUSD, providerName } from 'utils/web3'
 import withRpcProvider from 'hoc/withRpcProvider'
 import BuySellModal from 'components/buySell/BuySellModal'
+import { isMobileMetamask } from 'utils/device'
 
 import mixpanel from 'utils/mixpanel'
 import { truncateDecimals } from '../../utils/math'
@@ -181,6 +182,22 @@ const BuySellWidget = ({
     }
   }
 
+  /* Mobile Metamask app has this bug where it doesn't throw an exception on contract
+   * call when user rejects the transaction. Interestingly if you quit and re-enter
+   * the app after you reject the transaction the correct error with "user rejected..."
+   * message is thrown.
+   *
+   * As a workaround we hide the "waiting for user" modal after 5 seconds no matter what the
+   * user does if environment is the mobile metamask.
+   */
+  const mobileMetamaskHack = (prependStage) => {
+    if (isMobileMetamask()) {
+      setTimeout(() => {
+        setBuyWidgetState(`${prependStage}buy`)
+      }, 5000)
+    }
+  }
+
   const onMintOusd = async (prependStage) => {
     const mintedCoins = []
     setBuyWidgetState(`${prependStage}waiting-user`)
@@ -239,6 +256,7 @@ const BuySellWidget = ({
       }
 
       let result
+      mobileMetamaskHack(prependStage)
       if (mintAddresses.length === 1) {
         result = await vaultContract.mint(mintAddresses[0], mintAmounts[0], {
           gasLimit,
@@ -270,6 +288,7 @@ const BuySellWidget = ({
       if (e.code !== 4001) {
         await storeTransactionError(`mint`, mintedCoins.join(','))
       }
+
       onMintingError(e)
       console.error('Error minting ousd! ', e)
       mixpanel.track('Mint tx failed', {
