+++ dapp/src/components/buySell/SellWidget.js
@@ -14,6 +14,7 @@ import AccountStore from 'stores/AccountStore'
 import AnimatedOusdStore from 'stores/AnimatedOusdStore'
 import DisclaimerTooltip from 'components/buySell/DisclaimerTooltip'
 import { gasLimits } from 'constants/Contract'
+import { isMobileMetamask } from 'utils/device'
 
 import mixpanel from 'utils/mixpanel'
 
@@ -113,6 +114,22 @@ const SellWidget = ({
     // can not include the `calculateSplits` call here because if would be too many contract calls
   }
 
+  /* Mobile Metamask app has this bug where it doesn't throw an exception on contract
+   * call when user rejects the transaction. Interestingly if you quit and re-enter
+   * the app after you reject the transaction the correct error with "user rejected..."
+   * message is thrown.
+   *
+   * As a workaround we hide the "waiting for user" modal after 5 seconds no matter what the
+   * user does if environment is the mobile metamask.
+   */
+  const mobileMetamaskHack = () => {
+    if (isMobileMetamask()) {
+      setTimeout(() => {
+        setSellWidgetState('sell now')
+      }, 5000)
+    }
+  }
+
   const onSellNow = async (e) => {
     mixpanel.track('Sell now clicked')
     const returnedCoins = positiveCoinSplitCurrencies.join(',')
@@ -145,6 +162,7 @@ const SellWidget = ({
     setSellWidgetState('waiting-user')
     if (sellAllActive || forceSellAll) {
       try {
+        mobileMetamaskHack()
         const result = await vaultContract.redeemAll({
           gasLimit: gasLimits.REDEEM_GAS_LIMIT,
         })
@@ -162,6 +180,7 @@ const SellWidget = ({
       }
     } else {
       try {
+        mobileMetamaskHack()
         const result = await vaultContract.redeem(
           ethers.utils.parseUnits(
             ousdToSell.toString(),
